Confirmation dialog

+++ mainwindow.cpp
@@ -151,9 +151,14 @@ bool MainWindow::uninstall(const QString &names, const QString &postuninstall)
     ui->tabWidget->setCurrentWidget(ui->tabOutput);
 
     lock_file->unlock();
+    bool success;
+    //simulate install of selections and present for confirmation
+    //if user selects cancel, break routine but return success to avoid error message
+    if (!confirmActions(names, "remove")) return true;
+
     ui->tabWidget->setTabText(ui->tabWidget->indexOf(ui->tabOutput), tr("Uninstalling packages..."));
     displayOutput();
-    bool success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 remove " + names);
+    success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 remove -y " + names); //use -y since there is a confirm dialog already
     if (!postuninstall.isEmpty()) {
         qDebug() << "Post-uninstall";
         ui->tabWidget->setTabText(ui->tabWidget->indexOf(ui->tabOutput), tr("Running post-uninstall operations..."));
@@ -336,7 +341,7 @@ QString MainWindow::getLocalizedName(const QDomElement element) const
 {
     // pass one, find fully localized string, e.g. "pt_BR"
     QDomElement child = element.firstChildElement();
-    for (; (!child.isNull()); child = child.nextSiblingElement() ) {
+    for (; (!child.isNull()); child = child.nextSiblingElement()) {
         if (child.tagName() == locale.name() && !child.text().trimmed().isEmpty()) {
             return child.text().trimmed();
         }
@@ -367,7 +372,7 @@ QString MainWindow::getLocalizedName(const QDomElement element) const
 // get translation for the category
 QString MainWindow::getTranslation(const QString item)
 {
-    if (locale.name() == "en_US" ) { // no need for translation
+    if (locale.name() == "en_US") { // no need for translation
         return item;
     }
 
@@ -936,6 +941,82 @@ void MainWindow::listFlatpakRemotes()
     ui->comboRemote->blockSignals(false);
 }
 
+// Display warning for Debian Backports
+bool MainWindow::confirmActions(QString names, QString action)
+{
+    qDebug() << "+++" << __PRETTY_FUNCTION__ << "+++";
+    qDebug() << "names" << names << "and" << change_list;
+    QString msg;
+
+    QString detailed_names;
+    QStringList detailed_installed_names;
+    QString detailed_to_install;
+    QString detailed_removed_names;
+    QString recommends;
+    if (tree == ui->treeFlatpak){
+        detailed_installed_names = change_list;
+    } else if (tree == ui->treeBackports) {
+        recommends = (ui->checkBoxInstallRecommendsMXBP->isChecked()) ? "--install-recommends " : "";
+        detailed_names = cmd.getCmdOut("DEBIAN_FRONTEND=gnome LANG=C apt-get -s -V -o=Dpkg::Use-Pty=0 " + action + " " + recommends + "-t " + ver_name + "-backports --reinstall " + names + "|grep 'Inst\\|Remv' | awk '{print $2 \";\" $3 \";\" $1}'");
+    } else if (tree == ui->treeMXtest) {
+        recommends = (ui->checkBoxInstallRecommendsMX->isChecked()) ? "--install-recommends " : "";
+        detailed_names = cmd.getCmdOut("DEBIAN_FRONTEND=gnome LANG=C apt-get -s -V -o=Dpkg::Use-Pty=0 " + action + " -t a=mx,c=test " + recommends +  "--reinstall " + names + "|grep 'Inst\\|Remv' | awk '{print $2 \";\" $3 \";\" $1}'");
+    } else {
+        recommends = (ui->checkBoxInstallRecommends->isChecked()) ? "--install-recommends " : "";
+        detailed_names = cmd.getCmdOut("DEBIAN_FRONTEND=gnome LANG=C apt-get -s -V -o=Dpkg::Use-Pty=0 " + action + " " + recommends +  "--reinstall " + names + "|grep 'Inst\\|Remv'| awk '{print $2 \";\" $3 \";\" $1}'");
+    }
+
+    if (tree != ui->treeFlatpak){
+        detailed_installed_names=detailed_names.split("\n");
+    }
+
+    detailed_installed_names.sort();
+    qDebug() << "detailed installed names sorted " << detailed_installed_names;
+    QStringListIterator iterator(detailed_installed_names);
+    if (tree != ui->treeFlatpak){
+        while (iterator.hasNext()){
+            QString value = iterator.next();
+            if (value.contains("Remv")){
+                value = value.section(";",0,0) + " " + value.section(";",1,1);
+                detailed_removed_names = detailed_removed_names + value + "\n";
+            }
+            if (value.contains("Inst")){
+                value = value.section(";",0,0) + " " + value.section(";",1,1) + ")";
+                detailed_to_install = detailed_to_install + value + "\n";
+            }
+        }
+        if (!detailed_removed_names.isEmpty()){
+            detailed_removed_names.prepend(tr("Remove") + "\n");
+        }
+        if (!detailed_to_install.isEmpty())
+            detailed_to_install.prepend(tr("Install") + "\n");
+    } else {
+        if (action == "remove") {
+            detailed_removed_names = change_list.join("\n");
+            detailed_to_install.clear();        }
+        if (action == "install") {
+            detailed_to_install = change_list.join("\n");
+            detailed_removed_names.clear();
+        }
+    }
+
+        msg = "<b>" + tr("The following packages were selected. Click Show Details for list of changes.") + "</b>";
+
+        QMessageBox msgBox;
+        msgBox.setText(msg);
+        msgBox.setInformativeText("\n" + names);
+        msgBox.addButton(QMessageBox::Ok);
+        msgBox.addButton(QMessageBox::Cancel);
+        if (action == "install"){
+            //msgBox.setDetailedText(tr("Install") + "\n" + detailed_names + "\n" + tr("Remove") + "\n" + detailed_names);
+            msgBox.setDetailedText(detailed_to_install + "\n" + detailed_removed_names);
+        } else {
+            msgBox.setDetailedText(detailed_removed_names + "\n" + detailed_to_install);
+        }
+        msgBox.setFixedWidth(700);
+        return msgBox.exec() == QMessageBox::Ok;
+}
+
 // Install the list of apps
 bool MainWindow::install(const QString &names)
 {
@@ -953,16 +1034,21 @@ bool MainWindow::install(const QString &names)
     bool success = false;
 
     QString recommends;
+
+    //simulate install of selections and present for confirmation
+    //if user selects cancel, break routine but return success to avoid error message
+    if (!confirmActions(names, "install")) return true;
+
     displayOutput();
     if (tree == ui->treeBackports) {
         recommends = (ui->checkBoxInstallRecommendsMXBP->isChecked()) ? "--install-recommends " : "";
-        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install " + recommends +  "-t " + ver_name + "-backports --reinstall " + names);
+        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install -y " + recommends +  "-t " + ver_name + "-backports --reinstall " + names);
     } else if (tree == ui->treeMXtest) {
         recommends = (ui->checkBoxInstallRecommendsMX->isChecked()) ? "--install-recommends " : "";
-        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install " + recommends +  "-t a=mx,c=test " + names);
+        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install -y " + recommends +  "-t a=mx,c=test " + names);
     } else {
         recommends = (ui->checkBoxInstallRecommends->isChecked()) ? "--install-recommends " : "";
-        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install " + recommends +  "--reinstall " + names);
+        success = cmd.run("DEBIAN_FRONTEND=gnome apt-get -o=Dpkg::Use-Pty=0 install -y " + recommends +  "--reinstall " + names);
     }
     lock_file->lock();
     bar->setValue(bar->maximum());
@@ -1358,7 +1444,7 @@ void MainWindow::clearUi()
         ui->labelNumInstMX->clear();
         ui->labelNumUpgrMX->clear();
         ui->treeMXtest->clear();
-    } else if (tree == ui->treeBackports || tree == ui->treePopularApps ) {
+    } else if (tree == ui->treeBackports || tree == ui->treePopularApps) {
         ui->labelNumApps_3->clear();
         ui->labelNumInstBP->clear();
         ui->labelNumUpgrBP->clear();
@@ -1562,7 +1648,7 @@ QHash<QString, VersionNumber> MainWindow::listInstalledVersions()
 {
     qDebug() << "+++" << __PRETTY_FUNCTION__ << "+++";
     disconnect(conn);
-    QString out = cmd.getCmdOut("dpkg -l | grep '^ii'");
+    QString out = cmd.getCmdOut("dpkg -l | grep '^ii'", true);
     conn = connect(&cmd, &Cmd::outputAvailable, [](const QString &out) { qDebug() << out.trimmed(); });
 
     QString name;
@@ -1721,7 +1807,7 @@ void MainWindow::findPopular() const
             item->setExpanded(true);
             item->setHidden(false);
             int count = item->childCount();
-            for (int i = 0; i < count; ++i ) {
+            for (int i = 0; i < count; ++i) {
                 item->child(i)->setHidden(false);
             }
         }
@@ -1780,7 +1866,7 @@ void MainWindow::showOutput()
 void MainWindow::on_buttonInstall_clicked()
 {
     qDebug() << "+++" << __PRETTY_FUNCTION__ << "+++";
-
+    //qDebug() << "change list"  << .join(" ");
     showOutput();
 
     if (tree == ui->treePopularApps) {
@@ -1797,6 +1883,18 @@ void MainWindow::on_buttonInstall_clicked()
             QMessageBox::critical(this, tr("Error"), tr("Problem detected while installing, please inspect the console output."));
         }
     } else if (tree == ui->treeFlatpak) {
+        //confirmation dialog
+        if (!confirmActions(change_list.join(" "), "install")) {
+            displayFlatpaks(true);
+            indexFilterFP.clear();
+            ui->comboFilterFlatpak->setCurrentIndex(0);
+            QMessageBox::information(this, tr("Done"), tr("Processing finished successfully."));
+            ui->tabWidget->blockSignals(true);
+            ui->tabWidget->setCurrentWidget(ui->tabFlatpak);
+            ui->tabWidget->blockSignals(false);
+            enableTabs(true);
+            return;
+        }
         setCursor(QCursor(Qt::BusyCursor));
         displayOutput();
         if (cmd.run("runuser -l $(logname) -c \"socat SYSTEM:'flatpak install -y " + user + ui->comboRemote->currentText() + " " + change_list.join(" ") + "',stderr STDIO\"")) {
@@ -1926,6 +2024,21 @@ void MainWindow::on_buttonUninstall_clicked()
         QString conf = "-y ";
         if (fp_ver  < VersionNumber("1.0.1")) {
             conf = "";
+        }
+        //confirmation dialog
+        if (!confirmActions(change_list.join(" "), "remove")) {
+            displayFlatpaks(true);
+            indexFilterFP.clear();
+            listFlatpakRemotes();
+            ui->comboRemote->setCurrentIndex(0);
+            on_comboRemote_activated(ui->comboRemote->currentIndex());
+            ui->comboFilterFlatpak->setCurrentIndex(0);
+            QMessageBox::information(this, tr("Done"), tr("Processing finished successfully."));
+            ui->tabWidget->blockSignals(true);
+            ui->tabWidget->setCurrentWidget(ui->tabFlatpak);
+            ui->tabWidget->blockSignals(false);
+            enableTabs(true);
+            return;
         }
 
         setCursor(QCursor(Qt::BusyCursor));
@@ -2172,7 +2285,7 @@ void MainWindow::filterChanged(const QString &arg1)
             found_items = tree->findItems("not installed", Qt::MatchExactly, 5);
             ui->labelNumAppFP->setText(QString::number(found_items.count()));
             while (*it) {
-                if (found_items.contains(*it) ) {
+                if (found_items.contains(*it)) {
                     (*it)->setHidden(false);
                     (*it)->setText(6, "true"); // Displayed flag
                 } else {
@@ -2210,7 +2323,7 @@ void MainWindow::filterChanged(const QString &arg1)
     }
 
     while (*it) {
-        if (found_items.contains(*it) ) {
+        if (found_items.contains(*it)) {
             (*it)->setHidden(false);
             (*it)->setText(6, "true"); // Displayed flag
         } else {
